Refactor: move HunyuanVideo-specific defaults out of NetworkTrainer base
trainer_base.py had HunyuanVideo-specific implementations (load_vae,
load_transformer, call_dit, process_sample_prompts, do_inference, ...) as
default methods, leaving hv_train_network.py with only parser + main while
other architectures properly subclass NetworkTrainer. This broke the symmetry
and forced HV-specific imports (HYVideoDiffusionTransformer, VAE_VER,
text_encoder_module, ...) into the shared base file.

Introduce HunyuanVideoNetworkTrainer in hv_train_network.py and make the
corresponding methods in NetworkTrainer abstract hooks (NotImplementedError).
The base class is now architecture-agnostic, and HV becomes consistent with
Wan/Qwen/FLUX/FramePack/Z-Image/HV1.5/Kandinsky5 subclasses.

Completes the outstanding Step 1 item from references/refactoring_plan.md.
